feat: consolidate title bar controls into a single menu
The settings dropdown lived in the bottom control panel as ProfileGroup,
competing for space with the controls that actually run an interview, and
the title bar spent permanent width on three icon buttons - stealth, theme,
documentation - for actions used occasionally.

Both now live behind one menu button placed immediately before minimize.
The menu owns the dialogs it opens, so ConfigurationDialog moves off
MainPage and DocumentationDialog off Titlebar.

Titlebar renders outside RouterProvider, so Buy Credits navigates through
the router instance rather than useNavigate. Theme and documentation stay
reachable when logged out; the account items stay disabled unless the
assistant is idle, as they were before.

Start/Stop now sits between equal-basis side columns, which keeps it on the
panel's centre line whatever the audio, LLM, tools and zoom groups hold.

Closes #69
